+++ bukkit/src/main/java/com/github/games647/fastlogin/bukkit/listener/ProtcolSupportListener.java
@@ -8,6 +8,7 @@ import java.net.InetSocketAddress;
 import java.util.logging.Level;
 
 import org.bukkit.Bukkit;
+import org.bukkit.ChatColor;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
@@ -15,6 +16,7 @@ import org.bukkit.event.player.PlayerJoinEvent;
 
 import protocolsupport.api.events.PlayerLoginStartEvent;
 import protocolsupport.api.events.PlayerPropertiesResolveEvent;
+import protocolsupport.api.events.PlayerPropertiesResolveEvent.ProfileProperty;
 
 public class ProtcolSupportListener implements Listener {
 
@@ -34,12 +36,11 @@ public class ProtcolSupportListener implements Listener {
 
         String playerName = loginStartEvent.getName();
         if (plugin.getEnabledPremium().contains(playerName)) {
-            loginStartEvent.setOnlineMode(true);
-            InetSocketAddress address = loginStartEvent.getAddress();
-
-            PlayerSession playerSession = new PlayerSession(playerName, null, null);
-            plugin.getSessions().put(address.toString(), playerSession);
-//            loginStartEvent.setUseOnlineModeUUID(true);
+            //the player have to be registered in order to invoke the command
+            startPremiumSession(playerName, loginStartEvent, true);
+        } else if (plugin.getConfig().getBoolean("autologin") && !plugin.getAuthPlugin().isRegistered(playerName)) {
+            startPremiumSession(playerName, loginStartEvent, false);
+            plugin.getEnabledPremium().add(playerName);
         }
     }
 
@@ -50,8 +51,7 @@ public class ProtcolSupportListener implements Listener {
         if (session != null) {
             session.setVerified(true);
 
-            PlayerPropertiesResolveEvent.ProfileProperty skinProperty = propertiesResolveEvent.getProperties()
-                    .get("textures");
+            ProfileProperty skinProperty = propertiesResolveEvent.getProperties().get("textures");
             if (skinProperty != null) {
                 WrappedSignedProperty signedProperty = WrappedSignedProperty
                         .fromValues(skinProperty.getName(), skinProperty.getValue(), skinProperty.getSignature());
@@ -74,12 +74,35 @@ public class ProtcolSupportListener implements Listener {
                 if (player.isOnline()) {
                     //check if it's the same player as we checked before
                     if (session != null && player.getName().equals(session.getUsername()) && session.isVerified()) {
-                        plugin.getLogger().log(Level.FINE, "Logging player {0} in", player.getName());
-                        plugin.getAuthPlugin().forceLogin(player);
+                        if (session.needsRegistration()) {
+                            plugin.getLogger().log(Level.FINE, "Register player {0}", player.getName());
+
+                            String generatedPassword = plugin.generateStringPassword();
+                            plugin.getAuthPlugin().forceRegister(player, generatedPassword);
+                            player.sendMessage(ChatColor.DARK_GREEN + "Auto registered with password: "
+                                    + generatedPassword);
+                            player.sendMessage(ChatColor.DARK_GREEN + "You may want change it?");
+                        } else {
+                            plugin.getLogger().log(Level.FINE, "Logging player {0} in", player.getName());
+                            plugin.getAuthPlugin().forceLogin(player);
+                            player.sendMessage(ChatColor.DARK_GREEN + "Auto logged in");
+                        }
                     }
                 }
             }
             //Wait before auth plugin and we received a message from BungeeCord initializes the player
         }, DELAY_LOGIN);
+    }
+
+    private void startPremiumSession(String playerName, PlayerLoginStartEvent loginStartEvent, boolean registered) {
+        if (plugin.getApiConnector().isPremiumName(playerName)) {
+            loginStartEvent.setOnlineMode(true);
+            InetSocketAddress address = loginStartEvent.getAddress();
+
+            PlayerSession playerSession = new PlayerSession(playerName, null, null);
+            playerSession.setRegistered(registered);
+            plugin.getSessions().put(address.toString(), playerSession);
+//            loginStartEvent.setUseOnlineModeUUID(true);
+        }
     }
 }